requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
ipykernel
langchain_core
streamlit
langchain-groq